Extracted display of regions into a RegionCard component inside RegionInfoCard.
