+++ src/components/RegionInfoCard.tsx
@@ -1,8 +1,30 @@
-import {Box, Card, CardBody, Heading, Text} from 'grommet'
+import {Box, Card, CardBody, CardHeader, Heading, Text} from 'grommet'
 import Link from "next/link";
 import {useAtomValue} from "jotai";
 
 import {regionAssociationsAtom} from "../state/global";
+import {RegionNode} from "../utils/types";
+
+const oneEarthLabels = ['Realm','Subrealm','Bioregion','Ecoregion']
+const epaLabels = ['Level I','Level II','Level III','Level IV']
+
+const RegionCard = ({region, title}: {region: RegionNode, title: string}) => {
+    return (
+        <Card pad="small">
+            <CardHeader>{title}</CardHeader>
+            <CardBody margin={{top: "xxsmall"}}>
+                {region && (
+                    <Text>
+                        <LinkWrapper condition={region.link} url={region.link}>{region.name}</LinkWrapper>
+                    </Text>
+                )}
+            </CardBody>
+        </Card>
+    )
+}
+
+const LinkWrapper = ({ condition, url, children }: {condition: any, url: string | undefined, children: any}) =>
+    (condition && url) ? <Link href={url} target="_blank">{children}</Link> : children;
 
 export default function RegionInfoCard() {
     const associations = useAtomValue(regionAssociationsAtom)
@@ -22,15 +44,13 @@ export default function RegionInfoCard() {
                         {associations?.oneEarth && associations.oneEarth.length > 0 && (
                             <Box direction="column">
                                 <Heading level="4" margin={{vertical: "small"}}>One Earth</Heading>
-                                <Box direction="row">
-                                    <Box basis="1/2">
-                                        <Text>Realm: {associations.oneEarth[0] && (<Link
-                                            href={associations.oneEarth[0]?.link || ''}>{associations.oneEarth[0]?.name}</Link>)}</Text>
-                                    </Box>
-                                    <Box basis="1/2">
-                                        <Text>Bioregion: {associations.oneEarth[2] && (<Link
-                                            href={associations.oneEarth[2].link || ''}>{associations.oneEarth[2].name}</Link>)}</Text>
-                                    </Box>
+                                <Box direction="row" width="100%" gap="small">
+                                    {oneEarthLabels.map(
+                                        (label: string, index) => {
+                                            if (associations?.oneEarth && associations.oneEarth[index])
+                                                return <RegionCard key={index} region={associations.oneEarth[index]} title={label}/>
+                                        }
+                                    )}
                                 </Box>
                             </Box>
                         )}
@@ -38,13 +58,13 @@ export default function RegionInfoCard() {
                             <Box direction="column">
                                 <Heading level="4" margin={{vertical: "small"}}>EPA</Heading>
                                 <Box direction="row">
-                                    <Box basis="1/2">
-                                        <Text>Level 1: {associations.epa[0] && (<Link
-                                            href={associations.epa[0]?.link || ''}>{associations.epa[0]?.name}</Link>)}</Text>
-                                    </Box>
-                                    <Box basis="1/2">
-                                        <Text>Level 2: {associations.epa[1] && (<Link
-                                            href={associations.epa[1]?.link || ''}>{associations.epa[1]?.name}</Link>)}</Text>
+                                    <Box direction="row" width="100%" gap="small">
+                                        {epaLabels.map(
+                                            (label: string, index) => {
+                                                if (associations?.epa && associations.epa[index])
+                                                    return <RegionCard key={index} region={associations.epa[index]} title={label}/>
+                                            }
+                                        )}
                                     </Box>
                                 </Box>
                             </Box>
